[MODIFY] some quality changes
